Introduce quic_proto_dissector_dg_table
With it, the http3 dissector can dissect session_id from datagram frames.

(A rejected alternative approach: use the existing
quic_proto_dissector_table, and modify the data param (currently
stream_info) of call_dissector_with_data to contain whether the call
is about a stream or a datagram.  But this approach is not backward
compantible, and requires to update not just packet-http3, but
packet-nbt.c and packet-dns.c as well.